fix deepep moe for rocm
